fixed relu derivative, improved model 00 and 00b, added model 02d

+++ network/activation.py
@@ -49,7 +49,7 @@ class __ReLU(Activation):
         return np.maximum(x, 0, x)
 
     def backward(self, x: np.ndarray) -> np.ndarray:
-        return (x > 0).astype(int)
+        return x > 0
 
 
 @jitclass([])
